// src/enums.ts
export enum AppConsts {
    EMPTY = 0,
    FULL = 1,
    WIDTH = window.innerWidth - 30,
    HEIGHT = window.innerHeight - 20,
    BACKGROUNDCOLOR = 0x1099bb,
    MARINA_COLOR = 0xFFFF00,
    EMPTY_SHIP_COLOR = 0x86F442,
    FULL_SHIP_COLOR = 0xFF3300,
    SHIP_LOCATION_X = window.innerWidth + 150,
    SHIP_LOCATION_Y = window.innerHeight / 2 + 50,
    GATES_LOCATION_X = window.innerWidth / 4 + 50,
    LOWER_GATE_LOCATION_Y = window.innerHeight - window.innerHeight / 3,
    TIME_INTERVAL = 8000
}